Implemented gettign assemblies in the AssembliesHelper class
